refactor: get rid of services and types folders
