+++ src/server/features/hover.ts
@@ -2,10 +2,10 @@ import { tokenizer } from "scss-symbols-parser";
 import { MarkupKind, SymbolKind } from "vscode-languageserver";
 import type { Hover, MarkupContent } from "vscode-languageserver";
 import type { TextDocument } from "vscode-languageserver-textdocument";
+import { NodeType } from "../nodes";
 import { sassBuiltInModules } from "../sass-built-in-modules";
 import { sassDocAnnotations } from "../sassdoc-annotations";
-import type StorageService from "../services/storage";
-import { NodeType } from "../types/nodes";
+import type StorageService from "../storage";
 import type {
 	IScssDocument,
 	ScssSymbol,
@@ -14,8 +14,8 @@ import type {
 	ScssFunction,
 	ScssImport,
 	ScssForward,
-} from "../types/symbols";
-import type { Token } from "../types/tokens";
+} from "../symbols";
+import type { Token } from "../tokens";
 import { applySassDoc } from "../utils/sassdoc";
 import { asDollarlessVariable, getLimitedString } from "../utils/string";
