+++ src/server/features/signature-help.ts
@@ -3,14 +3,14 @@ import { MarkupKind, SignatureInformation } from "vscode-languageserver";
 import type { SignatureHelp } from "vscode-languageserver";
 import type { TextDocument } from "vscode-languageserver-textdocument";
 import { sassBuiltInModules } from "../sass-built-in-modules";
-import type StorageService from "../services/storage";
+import type StorageService from "../storage";
 import type {
 	IScssDocument,
 	ScssForward,
 	ScssFunction,
 	ScssImport,
 	ScssMixin,
-} from "../types/symbols";
+} from "../symbols";
 import { hasInFacts } from "../utils/facts";
 import { applySassDoc } from "../utils/sassdoc";
 import { asDollarlessVariable, getTextBeforePosition } from "../utils/string";
