+++ src/server/symbols.ts
