+++ src/test/features/go-definition.spec.ts
@@ -3,7 +3,7 @@ import { SymbolKind } from "vscode-languageserver";
 import { TextDocument } from "vscode-languageserver-textdocument";
 import { ScssDocument } from "../../server/document";
 import { goDefinition } from "../../server/features/go-definition";
-import StorageService from "../../server/services/storage";
+import StorageService from "../../server/storage";
 import * as helpers from "../helpers";
 
 const storage = new StorageService();
