+++ src/test/features/workspace-symbol.spec.ts
@@ -3,7 +3,7 @@ import { SymbolKind } from "vscode-languageserver";
 import { TextDocument } from "vscode-languageserver-textdocument";
 import { ScssDocument } from "../../server/document";
 import { searchWorkspaceSymbol } from "../../server/features/workspace-symbol";
-import StorageService from "../../server/services/storage";
+import StorageService from "../../server/storage";
 
 const storage = new StorageService();
